# Update & patch VMA, and re-implement the small buffers optimization