… thirdparty/vulkan/patches/03-VMA-universal-pools.patch
… thirdparty/vulkan/vk_mem_alloc.h