Since VMA's _master_ branch doesn't feature the small buffers optimization that was being sketched in a feature branch, this PR reformulates how that is achieved, with some benefits.

The optimization consisted in having in VMA an additional list of blocks in the default pool, which would be used for allocations <= 4 KiB and which was configured to ignore the device's allocation granularity.

The new approach consists in letting additional pools in VMA contain allocations of multiple memory types (in contrast to the original inteded usage, which is to allow memory type specific pools). This comes with the advantage of allowing applications to have additional pools as flexible as the default one and scatter allocations among them at wish and convenience.

Therefore, the Vulkan RD code is changed to leverage such addition, bringing back what was happening in the aforementioned, now outdated VMA's feature branch. That is, Vulkan RD creates an additional pool to store there objects of a size below some threshold (the 4 KiB from the VMA experiment is used).

The other aspect of the old VMA optimization (namely, setting granularity to 1) can now be used in non-default pools via the `VMA_POOL_CREATE_IGNORE_BUFFER_IMAGE_GRANULARITY_BIT` that VMA provides out of the box (and it can be used across memory types thanks the custom changes in this PR). However, my benchmarking has shown that the granularity trick wasn't involved in the performance improvement (and, as far as I can tell, would break correctness, so this PR is also an improvement in that regard, by not enabling it).

---

I've done some benchmarking in `debug_release` with an updated version of the script used in #51524 ([ClassProfiling4.zip](https://github.com/godotengine/godot/files/8051871/ClassProfiling4.zip)), with the following results:

| # | Version | Benchmark time |
|--|--|--|
| 1 | Current _master_ | 6511 ms |
| 2 | _master_ + new VMA with no extra pool | 6549 ms |
| 3 | _master_ + new VMA with small objects pool | 5035 ms | <- This PR
| 4 | _master_ + new VMA with small objects pool + granularity hack | 5125 ms |

Being 1 the baseline, I'd expect 2 to be slower, but maybe the new VMA already does a good job without any optimization effort. In any case, the small objects pool added to Vulkan RD seems to be beneficial as shown in 3. I would've expected 4 to be faster than 3, but, given the correctness concerns, we don't really care (we even get a greater peace of mind because there's no temptation to enable the granularity trick with fingers crossed).

With all that said, I believe this PR improves the state of things, but I'm not sure how useful it will be for #48978.

P. S.: If the changes done to VMA here for having universal pools are wanted upstream, I am willing to make it opt-in (via a creation flag or something) and submit a PR to their repo so users can benefit from the new pool semantics if they need them while keeping the current behavior still available.

P. P. S.: An additional benefit of this approach for Godot and other applications is that the allocation algorithm for the small objects pool can be picked separately. For now, I've left the linear one for Godot, as that's the one used before this PR, but we may do some experimentation to maybe find a better option.