@akien-mga 2022-02-12
I'm not familiar with those but Clang seems to raise some warnings about nullability completeness: https://github.com/godotengine/godot/runs/5162802525?check_suite_focus=true