@RandomShaper 2022-02-12
Looks like an upstream issue. I'll make sure and write a patch.