@RandomShaper 2022-02-12
Fixed nullability errors, and upstreamed: https://github.com/GPUOpen-LibrariesAndSDKs/VulkanMemoryAllocator/pull/235